DocumentsUI: The title disappears after rotation
The onItemCheckedStateChanged() method in DirectoryFragment.java
called the setTitle() of ActionMode to modify the title of the
action bar. But this change is not redone after the activity is
recreated. Therefore, when the orientation is changed, the change
in title will be lost.

Set the android:configChanges for DocumentsActivity to avoid
recreating this activity.

CRs-Fixed: 641604

Change-Id: I768422b63a07fcb3ac2c8fa5c6007e49965ff713