Keep ADB_ENABLED database value in sync with actual adb state
This change will update the ADB_ENBLED property in database to
the actual adb configuration as set in UsbDeviceManager.
It takes care of setting ADB_ENBLED to 0 in Charging mode
as adb should be OFF in this mode.Also, When user changes from
charging to some other mode, it restores ADB_ENBLED to what it
was before user changed to Charging mode.

Change-Id: I1ff9b8602f58da9a47ebfb2684ca86ce604ba046