SystemUI: Do not set preferred APN when not found
SystemUI may get wrong MNC for 4G CDMA card, as it is not in phone
process. After settings set preferred APN, SystemUI could not find
the APN and then update the preferred APN to be a wrong one.

Let SystemUI only read the preferred APN, do not to set when not found.

Change-Id: I0737b775a4f598c81b5b25e3d7f848a9dc49aaec
CRs-Fixed: 685822