Ultrasound: Rotate DigitalPenOffScreenDisplay when main display is rotated
Main display may sometimes mirror coordinates received on
the DigitalPenOffScreenDisplay. When the main display is rotated
and the external display is not, the coordinate systems do not match and
the coordinates received seem off.
This change would rotate the external display whenever the main display
is rotated to address that issue.

Change-Id: I08a8359c70cd7966d6b0a981b7a2f24400ef671e